# api/swagger.yaml
swagger: "2.0"
info:
  description: "Im tired and doing @EnableSwagger2 literally breaks my program :)"
  version: "1.3.3.7"
  title: "Robins Pain"
host: "localhost:8080"
...
paths:
...
  /sign_up:
    post:
      tags:
      - "sign_up"
      summary: "Add a new user"
      description: "Adds the new user to the database"
      operationId: "saveUser"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User to be registered"
        required: true
        schema:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
      responses:
        "201":
          description: "Created"
        "400":
          description: "Bad request"
...